Cleaned up Executor Class
Added more initialization code in the constructor.  As well as looking
up the host name the constructor now handles creation of a local UDP
socket from which we'll be sending commands.  It also initializes our
local sockaddr_in structure (pointing to the remote socket that will
receive commands).  This is hard coded for IPv4 at the moment.  Might
make it handle v4/v6 based on the information returned by
gethostbyname().

The constructor also produces a preamble string containing all the
information that needs to be sent with each RCON command so that it
is not re-calculated it each time a command is sent.

Server Name, hostent, Password & Port are no longer stored in the class'
private data members.  Instead the class uses these during the
constructor and then saves the resulting objects (socket, sockaddr_in,
preamble string).