Minor re-factor of Executor Class
The public methods of this class had most of their code in common, which
has now been moved into a private m_send() method.  Thus the public
methods now build the 2nd part of the RCON command string which they
pass on to m_send().  m_send() then adds the preamble and sends the
resulting buffer to the remote server.